changed resource page
-added aws-sdk to package.json
-added moment to package.json
-changed time to use moment
-made profile mobile friendly
