+++ package.json
@@ -6,8 +6,10 @@
     "ant-design-pro": "^2.2.1",
     "antd": "^3.14.1",
     "aws-amplify": "^1.1.22",
+    "aws-sdk": "^2.449.0",
     "draft-js": "^0.10.5",
     "gg-editor": "^2.0.2",
+    "moment": "^2.24.0",
     "react": "^16.8.2",
     "react-dom": "^16.8.2",
     "react-markdown": "^4.0.6",
