+++ src/components/pages/main/Main.js
@@ -22,7 +22,7 @@ export class Main extends Component {
     super(props);
 
     this.state = {
-      component: 1,
+      component: 3,
       header: 'General',
       visible: false,
     };
