+++ src/components/pages/profile/ProfileFeed.js
@@ -61,6 +61,7 @@ export class ProfileFeed extends Component {
         ]
       })
     ));
+    
     this.setState({componentLoading: false});
   }
